Upd version

+++ BaseProject.DAO/Models/Views/ArquivoVM.cs
@@ -4,7 +4,7 @@ namespace BaseProject.DAO.Models.Views
 	{
 		public string Nome { get; set; }
 		public string Extensao { get; set; }
-		public int Tamanho { get; set; }
+		public long Tamanho { get; set; }
         public string Tipo { get; set; }
         public string Base64 { get; set; }
